feat: remove duplicate download logic
